requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asyncmy==0.2.9
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
passlib==1.7.4
pyasn1==0.6.0
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
